extend DomainLayer
- update RxRealm template
- update CombineRealm template
- update Network+Moya template
