--- Templates/File Template/Common Services/Base Network Layer.xctemplate/JSONResponse.swift
@@ -1,13 +1,0 @@
-//
-//  JSONResponse.swift
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
-import Foundation
-
-protocol JSONResponse {
-    var status: ResponseStatus { get set }
-}
