--- Templates/File Template/Common Services/Realm Combine Storage Manager.xctemplate/DatabaseError.swift
@@ -1,16 +1,0 @@
-//
-//  DatabaseError.swift
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
-import Foundation
-
-enum DatabaseError: Error {
-    case unexpectedError
-    case saveFailed(String)
-    case deleteFailed(String)
-    case objectNotExists
-}
