+++ Templates/File Template/Domain Layer/Network+Moya.xctemplate/APIService.swift
@@ -1,11 +1,3 @@
-//
-//  APIService.swift
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
 import Foundation
 import Moya
 /*
