+++ Templates/File Template/Domain Layer/Network+Moya.xctemplate/ErrorCode.swift
@@ -1,11 +1,3 @@
-//
-//  ErrorCode.swift
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
 import Foundation
 
 // MARK: - ErrorCode
