+++ Templates/File Template/Domain Layer/Network+Moya.xctemplate/JSONResponse.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol JSONResponse {
+    var status: ResponseStatus { get set }
+}
