+++ Templates/File Template/Domain Layer/Network+Moya.xctemplate/NetworkProvider.swift
@@ -1,15 +1,7 @@
-//
-//  NetworkProvider.swift
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
 import Moya
 import RxSwift
 
-class NetworkProvider {
+final class NetworkProvider {
     
     // - Private Properties
     private var provider: MoyaProvider<APIService>
