+++ Templates/File Template/Domain Layer/Network+Moya.xctemplate/ServerError.swift
@@ -1,11 +1,3 @@
-//
-//  ServerError.swift
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
 import Foundation
 
 // MARK: - ServerError
