+++ Templates/File Template/Domain Layer/Storage+CombineRealm.xctemplate/DatabaseError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public enum DatabaseError: Error {
+    case unexpectedError
+    case saveFailed(String)
+    case deleteFailed(String)
+    case objectNotExists
+}
