+++ Templates/File Template/Domain Layer/Storage+RxRealm/RealmMigrationService.swift
@@ -1,16 +1,11 @@
-//
-//  RealmMigrationService.swift
-//  ___PROJECTNAME___
-//
-//  Created by ___FULLUSERNAME___ on ___DATE___.
-//  ___COPYRIGHT___
-//
-
-
 import Foundation
 import RealmSwift
 
-class RealmMigrationService {
+public protocol RealmMigrationServiceType {
+    func performMigration()
+}
+
+public final class RealmMigrationService: RealmMigrationServiceType {
     
     private var schemaVersion: String? {
         return Bundle.main.infoDictionary?["CFBundleVersion"] as? String
